requirements.txt
aiohttp==3.5.4
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
discord==1.0.1
discord.py==1.2.5
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.4
pylint==2.4.4
python-dotenv==0.10.5
requests==2.22.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.7
websockets==6.0
wrapt==1.11.2
yarl==1.4.2
